LLM: Fix static analysis and compile warnings.
IncludeTool warning: we can not create a duplicate definition of PLATFORM_USES_FIXED_GMalloc_CLASS because some platforms are setting it in headers rather than c++.
Change the c-language includes of LowLevelMemTrackerDefines to be predicated only on LLM_ENABLED_IN_CONFIG to avoid the need for the definition of PLATFORM_USES_FIXED_GMalloc_CLASS.

Fix implicit conversion warnings in VirtualAllocPageStatus.cpp.

#rb None, fixing build break
#rnx
#review Danny.Couture
#preflight 6362f378522c8f7ab35ccee0

[CL 22930788 by Matt Peters in ue5-main branch]

+++ Engine/Source/Runtime/Core/Private/Windows/VirtualAllocPageStatus.cpp
@@ -143,10 +143,10 @@ uint64& FHashMapLinearProbingVAlloc::FindOrAdd(uint64 Key, uint64 ValueIfMissing
 {
 	LLMCheck(Key <= MaxKey);
 
-	if (NumActive > NumBuckets * DesiredPopulation)
+	if (NumActive > static_cast<int64>(NumBuckets * DesiredPopulation))
 	{
 		Realloc(NumBuckets * 2);
-		LLMCheck(NumActive < NumBuckets * DesiredPopulation);
+		LLMCheck(NumActive < static_cast<int64>(NumBuckets * DesiredPopulation));
 	}
 
 	int64 FirstBucketIndex = static_cast<int64>(Key & (NumBuckets - 1));
